LayoutTests/imported/w3c/web-platform-tests/workers/shared-worker-name-via-options-expected.txt

CONSOLE MESSAGE: line 16: ReferenceError: Can't find variable: SharedWorker

FAIL Test the name property of shared workers mixing constructor options and constructor strings ReferenceError: Can't find variable: SharedWorker
